requirements.txt
aws-cdk-lib~=2.176.0
constructs>=10.0.0
boto3>=1.28.57
botocore>=1.31.57
python-dotenv>=1.0.0